Don't assume config file extension
Even if we want to enforce one extension,
doing it in requires scattered across the whole project
doesn't sound like a good idea (don't repeat yourself).

This commit also enables .js configs,
so I added that to .gitignore.
I'll document it when we settle on default config format.

Why was .json everywhere in the first place?
I thought it was neccessary in some older versions of Node
to be able to require a .json file.